requirements.txt
numpy<2
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
plotly
